+++ Conservatorio.BL/Interfaces/IEstudiante.cs
@@ -8,6 +8,6 @@ namespace Conservatorio.BL.Interfaces
         List<Estudiante> ObtenerEstudiantes(string keyword);
         List<Estudiante> ObtenerEstudiantes();
         void CrearEstudiante(Estudiante estudiante);
-        void EliminarEstudiante(long idEstudiante);
+        void EliminarEstudiante(int idEstudiante);
     }
 }
